# jaxrs-server/src/gen/java/io/swagger/model/BookAndLoanViewModel.java
/*
 * Johanna's Library
 * This is a book repository for Jóhanna the húsfreyja at Efri-Brú
 *
 * OpenAPI spec version: 1.0.0
 * Contact: birkirb15@ru.is
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.Date;
import javax.validation.constraints.*;

/**
 * BookAndLoanViewModel
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2017-09-26T12:51:20.012Z")
public class BookAndLoanViewModel   {
  @JsonProperty("id")
  private Integer id = null;

  @JsonProperty("title")
  private String title = null;

  @JsonProperty("firstName")
  private String firstName = null;

  @JsonProperty("lastName")
  private String lastName = null;

  @JsonProperty("datePublished")
  private Date datePublished = null;

  @JsonProperty("isbn")
  private String isbn = null;

  @JsonProperty("dateLoaned")
  private Date dateLoaned = null;

  @JsonProperty("dateReturned")
  private Date dateReturned = null;

  public BookAndLoanViewModel id(Integer id) {
    this.id = id;
    return this;
  }

  /**
   * Get id
   * @return id
   **/
  @JsonProperty("id")
  @ApiModelProperty(value = "")
  public Integer getId() {
    return id;
  }

  public void setId(Integer id) {
    this.id = id;
  }

  public BookAndLoanViewModel title(String title) {
    this.title = title;
    return this;
  }

  /**
   * Get title
   * @return title
   **/
  @JsonProperty("title")
  @ApiModelProperty(value = "")
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public BookAndLoanViewModel firstName(String firstName) {
    this.firstName = firstName;
    return this;
  }

  /**
   * Get firstName
   * @return firstName
   **/
  @JsonProperty("firstName")
  @ApiModelProperty(value = "")
  public String getFirstName() {
    return firstName;
  }

  public void setFirstName(String firstName) {
    this.firstName = firstName;
  }

  public BookAndLoanViewModel lastName(String lastName) {
    this.lastName = lastName;
    return this;
  }

  /**
   * Get lastName
   * @return lastName
   **/
  @JsonProperty("lastName")
  @ApiModelProperty(value = "")
  public String getLastName() {
    return lastName;
  }

  public void setLastName(String lastName) {
    this.lastName = lastName;
  }

  public BookAndLoanViewModel datePublished(Date datePublished) {
    this.datePublished = datePublished;
    return this;
  }

  /**
   * Get datePublished
   * @return datePublished
   **/
  @JsonProperty("datePublished")
  @ApiModelProperty(value = "")
  public Date getDatePublished() {
    return datePublished;
  }

  public void setDatePublished(Date datePublished) {
    this.datePublished = datePublished;
  }

  public BookAndLoanViewModel isbn(String isbn) {
    this.isbn = isbn;
    return this;
  }

  /**
   * Get isbn
   * @return isbn
   **/
  @JsonProperty("isbn")
  @ApiModelProperty(value = "")
  public String getIsbn() {
    return isbn;
  }

  public void setIsbn(String isbn) {
    this.isbn = isbn;
  }

  public BookAndLoanViewModel dateLoaned(Date dateLoaned) {
    this.dateLoaned = dateLoaned;
    return this;
  }

  /**
   * Get dateLoaned
   * @return dateLoaned
   **/
  @JsonProperty("dateLoaned")
  @ApiModelProperty(value = "")
  public Date getDateLoaned() {
    return dateLoaned;
  }

  public void setDateLoaned(Date dateLoaned) {
    this.dateLoaned = dateLoaned;
  }

  public BookAndLoanViewModel dateReturned(Date dateReturned) {
    this.dateReturned = dateReturned;
    return this;
  }

  /**
   * Get dateReturned
   * @return dateReturned
   **/
  @JsonProperty("dateReturned")
  @ApiModelProperty(value = "")
  public Date getDateReturned() {
    return dateReturned;
  }

  public void setDateReturned(Date dateReturned) {
    this.dateReturned = dateReturned;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    BookAndLoanViewModel bookAndLoanViewModel = (BookAndLoanViewModel) o;
    return Objects.equals(this.id, bookAndLoanViewModel.id) &&
        Objects.equals(this.title, bookAndLoanViewModel.title) &&
        Objects.equals(this.firstName, bookAndLoanViewModel.firstName) &&
        Objects.equals(this.lastName, bookAndLoanViewModel.lastName) &&
        Objects.equals(this.datePublished, bookAndLoanViewModel.datePublished) &&
        Objects.equals(this.isbn, bookAndLoanViewModel.isbn) &&
        Objects.equals(this.dateLoaned, bookAndLoanViewModel.dateLoaned) &&
        Objects.equals(this.dateReturned, bookAndLoanViewModel.dateReturned);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, title, firstName, lastName, datePublished, isbn, dateLoaned, dateReturned);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class BookAndLoanViewModel {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    firstName: ").append(toIndentedString(firstName)).append("\n");
    sb.append("    lastName: ").append(toIndentedString(lastName)).append("\n");
    sb.append("    datePublished: ").append(toIndentedString(datePublished)).append("\n");
    sb.append("    isbn: ").append(toIndentedString(isbn)).append("\n");
    sb.append("    dateLoaned: ").append(toIndentedString(dateLoaned)).append("\n");
    sb.append("    dateReturned: ").append(toIndentedString(dateReturned)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}