# swagger.yaml
swagger: "2.0"
info:
  description: "This is a book repository for Jóhanna the húsfreyja at Efri-Brú"
  version: "1.0.0"
  title: "Johanna's Library"
  contact:
    email: "birkirb15@ru.is"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8080"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /users:
    post:
      tags:
        - "users"
      summary: "Add new user to the database"
      description: ""
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object, that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/AddUserViewModel"
      responses:
        405:
          description: "Invalid input"
    get:
      tags:
      - "users"
      summary: "Gets all users"
      description: ""
      operationId: "getAllUsers"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserViewModel"
  /users/{userId}:
    get:
      tags:
      - "users"
      summary: "Get user by user ID"
      description: ""
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "The ID of the user that needs to be fetched"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/UserViewModel"
        400:
          description: "Invalid user id supplied"
        404:
          description: "User not found"
    put:
      tags:
      - "users"
      summary: "Update user"
      description: ""
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "Id of the user that needs to be updated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/AddUserViewModel"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      tags:
      - "users"
      summary: "Delete user"
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "The Id of the user that needs to be deleted"
        required: true
        type: "integer"
        format: "int32"
      responses:
        400:
          description: "Invalid user id supplied"
        404:
          description: "User not found"
...
  /users/loans/{date}:
    get:
      tags:
      - "users"
      - "loans"
      summary: "Get list of users that have a borrowed book at specified date or today by default. Includes the list of books that the user has loaned."
      description: ""
      operationId: "getUserAndLoanedBooksByDate"
      produces:
      - "application/json"
      parameters:
      - name: "date"
        in: "path"
        description: "The date specified to look up users that have borrowed a book"
        required: true
        type: "string"
        format: "date"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserAndBookLoansViewModel"
        400:
          description: "Invalid date supplied"
        404:
          description: "No users found"
  /users/loans/{date}/monthOverdue:
    get:
      tags:
      - "users"
      - "loans"
      summary: "Get list of users that have a book borrowed for a month or more at given date, defaults to today"
      description: ""
      operationId: "getUsersWithMonthOverdueBooksByDate"
      produces:
      - "application/json"
      parameters:
      - name: "date"
        in: "path"
        description: "The date specified to look up users that have borrowed a book"
        required: true
        type: "string"
        format: "date"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserViewModel"
        400:
          description: "Invalid date supplied"
        404:
          description: "No users found"
...
  /users/{userId}/reviews:
    get:
      tags:
      - "book reviews"
      summary: "Gets all book reviews that belong to specific user"
      description: ""
      operationId: "getBookReviewsByUserId"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "The id of the user that the reviews belong to"
        required: true
        type: "number"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/BookAndBookReviewViewModel"
...
definitions:
  AddUserViewModel:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
  UserViewModel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      firstName:
        type: "string"
      lastName:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
...
  BookViewModel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      title:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      datePublished:
        type: "string"
        format: "date-time"
      isbn:
        type: "string"
  BookAndLoanViewModel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      title:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      datePublished:
        type: "string"
        format: "date-time"
      isbn:
        type: "string"
      dateLoaned:
        type: "string"
        format: "date-time"
      dateReturned:
        type: "string"
        format: "date-time"
...
  UserAndBookLoansViewModel:
    type: "object"
    description: "User and list of books"
    properties:
      user: {
        "$ref": "#/definitions/UserViewModel"
      }
      books: {
        type: "array",
        items: {
          "$ref": "#/definitions/BookAndLoanViewModel"
        }
      }
...
  BookAndBookReviewViewModel:
    type: "object"
    properties:
      book: {
        "$ref": "#/definitions/BookViewModel"
      }
      rating:
        type: "integer"
        format: "int32"
      review:
        type: "string"